fix(refiner): keep the attention mask bias f32 — f16 cast overflowed -1e9 to -inf (#84)
The stock (non-chunked) attention path cast its additive bias to the compute
dtype; in f16 the -1e9 mask constant overflows to -inf, so any fully-masked
row (e.g. a pad query whose only visible keys are pad — a plain leading-pad
batch) softmaxes to NaN. dot_product_attention adds the bias to its f32
logits and the QK einsum stays F16_F16_F32, so passing the bias in f32 keeps
the tensor-core path while matching the finite-mask contract the chunked
path already made.

New tests build genuinely fully-masked rows in explicit f16 and require
finite output from both attention paths, module-level and end-to-end; they
fail with NaN on the old code (stock path) and run on both test lanes.

+++ plan_a_model.py
@@ -72,7 +72,12 @@ class CausalAttention(nnx.Module):
             bias = jnp.where(causal, 0.0, -1e9)[None, None, :, :]   # [1, 1, s, s]
             if pad_bias is not None:
                 bias = bias + pad_bias                              # pad_bias [b, 1, 1, s]
-            out = jax.nn.dot_product_attention(q, k, v, bias=bias.astype(x.dtype))
+            # The bias must stay f32: cast to f16 turns -1e9 into -inf (f16 max
+            # ~65504), and a fully-masked row would softmax to NaN (#84).
+            # dot_product_attention adds the bias to its f32 logits, so f16
+            # q/k/v keep the tensor-core path — same contract as the chunked
+            # branch, whose bias also stays f32.
+            out = jax.nn.dot_product_attention(q, k, v, bias=bias)
         return self.o(out.reshape(b, s, d))
